[REFACTOR] #72 메인 홈 파일 분리
