+++ momogum/Home/MainHome/Model/HomeModel.swift
@@ -1,0 +1,38 @@
+//
+//  HomeModel.swift
+//  momogum
+//
+//  Created by 김윤진 on 2/9/25.
+//
+
+//import Foundation
+//
+//struct HomeAPIResponse: Codable {
+//    let isSuccess: Bool
+//    let code: String
+//    let message: String
+//    let result: MealDiaryResult
+//}
+//
+//struct MealDiaryResult: Codable {
+//    let viewMealDiaryResponseList: [MealDiary]
+//}
+//
+//struct MealDiary: Codable, Identifiable {
+//    let id: Int
+//    let foodImageURLs: [String]
+//    let userImageURL: String
+//    let foodCategory: String
+//    let keyWord: [String]
+//    let isRevisit: String
+//
+//    enum CodingKeys: String, CodingKey {
+//        case id = "mealDiaryId"
+//        case foodImageURLs
+//        case userImageURL
+//        case foodCategory
+//        case keyWord
+//        case isRevisit
+//    }
+//}
+
